requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
rembg==2.0.67
onnxruntime==1.22.1
Pillow==11.3.0
python-dotenv==1.1.1
numpy==2.2.6
